requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
celery-pool-asyncio==0.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
defusedxml==0.7.1
distro==1.9.0
Django==5.1.3
django-celery-beat==2.7.0
django-timezone-field==7.0
dnspython==2.7.0
eventlet==0.38.0
exceptiongroup==1.2.2
fastapi==0.115.5
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
jiter==0.7.1
kombu==5.4.2
Mako==1.3.6
MarkupSafe==3.0.2
multidict==6.1.0
openai==1.55.0
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.48
propcache==0.2.0
psycopg2==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
pytest-asyncio==0.24.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
redis==5.2.0
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.2
starlette==0.41.2
tomli==2.1.0
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
untangle==1.2.1
urllib3==2.2.3
uvicorn==0.32.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.17.2
